1.1

+++ 1.1/1.1.c
@@ -8,7 +8,7 @@
  * \return a.
  */
 
-double get_a(float x, float y, float z);
+float get_a(float x, float y, float z);
 
 /**
  * @brief Функция расчитывающая и выводящая b
@@ -18,26 +18,28 @@ double get_a(float x, float y, float z);
  * \return b.
  */
 
-double get_b(float x, float y, float z);
+float get_b(float x, float y, float z);
 
 int main()
 {
     const float x = 1.4;
     const float y = 3.1;
     const float z = 0.5;
-    double a = get_a(x, y, z); // a=0.870404
-    double b = get_b(x, y, z); // b=-50.893158
+    float a = get_a(x, y, z); // a=0.870404
+    float b = get_b(x, y, z); // b=-50.893158
     printf("a= %f\n", a);
     printf("b= %f\n", b);
     return 0;
 }
-double get_a(float x, float y, float z)
+
+float get_a(float x, float y, float z)
 {
     float a = sqrt(z * x * sin(2 * x) + sqrt(expl(-2 * x) * (x + y)));
     float c = expl(-2 * x);
     return a;
 }
-double get_b(float x, float y, float z)
+
+float get_b(float x, float y, float z)
 {
     float b = expl(2 * x) * log(z + x) - pow(y, 3 * x) * log(y - x);
     return b;
